+++ dolomite_engine/arguments.py
@@ -271,8 +271,6 @@ class DistributedArgs(BaseArgs):
     sequence_parallel: bool = False
     # pipeline parallel world size
     pipeline_parallel_world_size: int = 1
-    # data parallel world size
-    data_parallel_size: int | None = None
     # distributed timeout for NCCL in minutes
     timeout_minutes: int | None = None
     # fsdp algorithm
