+++ dolomite_engine/utils/__init__.py
@@ -35,7 +35,6 @@ from .yaml import load_yaml
 def init_distributed(
     tensor_parallel_world_size: int,
     pipeline_parallel_world_size: int,
-    data_parallel_size: int,
     data_parallel_replication_world_size: int,
     data_parallel_sharding_world_size: int,
     zero_stage: int,
@@ -47,7 +46,6 @@ def init_distributed(
     Args:
         tensor_parallel_world_size (int): tensor parallel size
         pipeline_parallel_world_size (int): pipeline parallel size
-        data_parallel_size (int): data parallel size
         data_parallel_replication_world_size (int): data parallel replication world size
         data_parallel_sharding_world_size (int): data parallel sharding world size
         zero_stage (int): zero stage
@@ -58,7 +56,6 @@ def init_distributed(
     process_group_manager = ProcessGroupManager(
         tensor_parallel_world_size=tensor_parallel_world_size,
         pipeline_parallel_world_size=pipeline_parallel_world_size,
-        data_parallel_size=data_parallel_size,
         data_parallel_replication_world_size=data_parallel_replication_world_size,
         data_parallel_sharding_world_size=data_parallel_sharding_world_size,
         zero_stage=zero_stage,
