+++ tests/hf_models/test_common.py
@@ -5,9 +5,6 @@
 import json
 import os
 import tempfile
-from itertools import product
-from typing import Any, Callable
-from unittest import TestCase, skipUnless
 
 import torch
 from torch.testing import assert_close
@@ -16,15 +13,13 @@ from transformers import AutoConfig, AutoModelForCausalLM
 from dolomite_engine import SafeTensorsWeightsManager
 from dolomite_engine.hf_models import CommonConfig, GPTDolomiteConfig, export_to_huggingface, import_from_huggingface
 
+from ..test_common import BaseTestCommons
+
 
 _RUN_SLOW = True if os.getenv("RUN_SLOW", "False").lower() in ["1", "true"] else False
 
 
-class TestCommons(TestCase):
-    @staticmethod
-    def get_all_devices() -> list[torch.device]:
-        return [torch.device("cpu"), torch.device("cuda")]
-
+class TestCommons(BaseTestCommons):
     @staticmethod
     def get_attention_head_types() -> list[str]:
         return ["mha", "mqa", "gqa"]
@@ -36,21 +31,6 @@ class TestCommons(TestCase):
     @staticmethod
     def get_position_embedding_types() -> list[str]:
         return ["learned_absolute", "rope"]
-
-    @staticmethod
-    def get_dtypes() -> list[torch.dtype]:
-        return [torch.float32, torch.float16, torch.bfloat16]
-
-    def make_args_matrix(*args_lists) -> list[Any]:
-        return [p for p in product(*args_lists)]
-
-    def skip_test_if_device_unavailable(self, device: torch.device) -> None:
-        # convert to str
-        if isinstance(device, torch.device):
-            device = device.type
-
-        if device == "cuda" and not torch.cuda.is_available():
-            self.skipTest("skipping test because CUDA is unavailable")
 
     def skip_test_if_layernorm_kernel_unavailable(self, device: torch.device, dtype: torch.dtype) -> None:
         # convert to str
@@ -329,7 +309,3 @@ class TestCommons(TestCase):
                 assert_close(x, y, rtol=rtol_bfloat16, atol=atol_bfloat16)
             else:
                 raise ValueError(f"unexpected dtype ({dtype})")
-
-    @staticmethod
-    def slow_test(func: Callable) -> Callable:
-        return skipUnless(_RUN_SLOW, "skipping slow test since RUN_SLOW=True is not set in the environment")(func)
