+++ tests/training/test_commons.py
@@ -3,13 +3,14 @@
 # **************************************************
 
 import os
-from unittest import TestCase
 
 from dolomite_engine.arguments import TrainingArgs
 from dolomite_engine.utils import load_yaml
 
+from ..test_common import BaseTestCommons
 
-class TestCommons(TestCase):
+
+class TestCommons(BaseTestCommons):
     @staticmethod
     def load_training_args_for_unit_tests(filename: str) -> TrainingArgs:
         filepath = os.path.join(os.path.dirname(__file__), filename)
